Ensure one primary term per generation
This commit adds assertions that we have one primary term per translog
generation (note that does not mean there is a one-to-one correspondence
between primary terms and generations, it can still be the case, and is
expected, that a single primary term spans multiple translog
generations). With the addition of these assertions, we fix a bug where
we were not rolling the generation on primary promotion, thereby
violating this invariant.