Undo some inadvertent variable renames in SQL
We recently renamed "metaData" to "metadata" in all of the places. We
were slightly aggressive in doing this rename in the SQL codebase where
"MetaData" is used by the JDK for its naming of these classes, and we
tried to keep SQL consistent with that, but inadvertently renamed a few
things that we should not have. Also, since we did not make these
renames in the 7.x branch, we should rename them back in the master
branch for reasons of consistency too.

Relates #54519