Align analytics feature set usage serialization
The serialization here is not aligned with that in master, which assumes
that a >= 7.8.0 node is only writing the stats map to the stream. We
align the serialization here by doing that if the stream is >=
7.8.0. Note than when receiving from such a node, we can not
re-construct the backing response field, so we set it to null. This is
okay, since we should never re-serialize such an instance back on a
stream to a < 7.8.0 node, which needs the response instance for
serialization.